+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - Add SKSE plugin
 - Remove **ConsoleUtil** as requirement
 - Remove **Spell Perk Item Distributor (SPID)** as requirement
+- Remove clone NPC option
+- Remove delete NPC option
 
 ## 1.4.0